libs/cli: clean up package (#7806)
